Refactor dashboard view into model and presenter
