+++ src/ui/views/classic/index.js
@@ -1,10 +1,17 @@
 import resolvers, { classicResolvers } from './resolvers.js';
+import { renderDashboard as baseRenderDashboard } from '../../dashboard.js';
+import dashboardPresenter from './dashboardPresenter.js';
 
 const classicView = {
   id: 'classic',
   name: 'Classic Dashboard',
   resolvers,
-  presenters: {}
+  presenters: {
+    dashboard: dashboardPresenter
+  },
+  renderDashboard(state, summary) {
+    baseRenderDashboard(state, summary, dashboardPresenter);
+  }
 };
 
 export { classicResolvers };
